+++ src/anomalib/metrics/__init__.py
@@ -11,7 +11,7 @@ from typing import Any
 import torchmetrics
 from omegaconf import DictConfig, ListConfig
 
-from . import per_image
+from . import pimo
 from .anomaly_score_distribution import AnomalyScoreDistribution
 from .aupr import AUPR
 from .aupro import AUPRO
@@ -20,7 +20,7 @@ from .collection import AnomalibMetricCollection
 from .f1_max import F1Max
 from .f1_score import F1Score
 from .min_max import MinMax
-from .per_image import AUPIMO, PIMO
+from .pimo import AUPIMO, PIMO
 from .precision_recall_curve import BinaryPrecisionRecallCurve
 from .pro import PRO
 from .threshold import F1AdaptiveThreshold, ManualThreshold
@@ -37,7 +37,7 @@ __all__ = [
     "ManualThreshold",
     "MinMax",
     "PRO",
-    "per_image",
+    "pimo",
     "PIMO",
     "AUPIMO",
 ]
