+++ src/anomalib/metrics/pimo/binclf_curve.py
