+++ src/anomalib/metrics/pimo/dataclasses.py
