+++ src/anomalib/metrics/pimo/enums.py
