+++ src/anomalib/metrics/pimo/functional.py
