+++ src/anomalib/metrics/pimo/pimo.py
