+++ tests/unit/metrics/pimo/test_pimo.py
@@ -13,8 +13,8 @@ import pytest
 import torch
 from torch import Tensor
 
-from anomalib.metrics.per_image import functional, pimo
-from anomalib.metrics.per_image.pimo import AUPIMOResult, PIMOResult
+from anomalib.metrics.pimo import functional, pimo
+from anomalib.metrics.pimo.pimo import AUPIMOResult, PIMOResult
 
 
 def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
